Bug: #5505 - DicomAttributeSL.GetUInt64 has an errant cast to a UInt16 '''Schema/Configuration Changes: N/A'''  * Fixed.

+++ Dicom/DicomAttributeBinary.cs
@@ -2475,7 +2475,7 @@ namespace ClearCanvas.Dicom
             }
             else
             {
-                value = (UInt16)_values[index];
+                value = (UInt64)_values[index];
                 return true;
             }
         }
